+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## staged
 
+- Fixed bug in center-tap transformer modeling
 - Add wye-connected CapControl for IVR and FOT (polar) formulations
 - Fixed indexing issue for single-phase delta load models in linear formulations (LinDist3Flow, FOTP, FOTR, FBS)
 - Added ZIP load model